+++ __tests__/setup.tsx
@@ -9,8 +9,15 @@ jest.mock('@web3-react/core');
 jest.mock('modules/web3/helpers/ens');
 
 jest.mock('remark-gfm', () => () => null);
-jest.mock('remark-html', () => () => null);
+jest.mock('remark-parse', () => () => null);
 jest.mock('remark', () => () => null);
+jest.mock('rehype-stringify', () => () => null);
+jest.mock('remark-rehype', () => () => null);
+jest.mock('rehype-sanitize', () => () => null);
+jest.mock('unified', () => () => null);
+
+
+
 
 jest.mock('modules/address/components/AddressIcon', () => {
   return {
