+++ modules/delegates/api/fetchGithubDelegates.ts
@@ -2,7 +2,7 @@ import matter from 'gray-matter';
 import { SupportedNetworks } from 'modules/web3/constants/networks';
 import { fsCacheGet, fsCacheSet } from 'lib/fscache';
 import { fetchGitHubPage, GithubPage } from 'lib/github';
-import { markdownToHtml } from 'lib/utils';
+import { markdownToHtml } from 'lib/markdown';
 import { DelegateRepoInformation } from 'modules/delegates/types';
 import { getDelegatesRepositoryInformation } from './getDelegatesRepositoryInfo';
