+++ package.json
@@ -79,7 +79,6 @@
     "recharts": "^2.1.2",
     "rehype-sanitize": "^5.0.1",
     "rehype-stringify": "^9.0.3",
-    "remark": "^14.0.2",
     "remark-gfm": "^3.0.1",
     "remark-html": "^15.0.1",
     "remark-parse": "^10.0.1",
